Move adapters to separate Adapters package/folder.
